+++ src/components/EmptyView/index.js
@@ -10,7 +10,7 @@ export default function renderEmpty(loader, renderEmpty) {
     } else if (loader.error) {
       return <div style={style}>出错啦!{loader.error.message}<span onClick={() => loader.refresh()}>点我重试</span></div>
     } else if (renderEmpty) {
-      return renderEmpty(loader)
+      return typeof renderEmpty === 'function' ? renderEmpty(loader) : renderEmpty
     } else {
       return <div style={style}>没有数据,<span onClick={() => loader.refresh()}>点我重试</span></div>
     }