+++ src/components/UserAreaView/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import { isPWA } from 'utils/utils'
+import { isPWAorMobile } from 'utils/utils'
 
 export default function UserAreaView({ children, className, style, bar }) {
   return (
     <div className="full-height">
-      {bar && isPWA() ? (
+      {bar && isPWAorMobile() ? (
         <div style={{ position: 'relative', zIndex: 8888 }}>
           <div style={{ height: 'env(safe-area-inset-top)' }}>
             <div style={{ position: 'fixed', right: 0, left: 0, top: 0, backgroud: '#000', height: 'env(safe-area-inset-top)', paddingTop: 'env(safe-area-inset-top)' }}></div>