+++ src/contexts/navi/index.js
@@ -1,12 +1,14 @@
 import React, { useContext as useReactContext } from 'react'
 import { MIconView } from 'components'
+import { useRouterContext } from '../router'
 // 上下文context.避免react多级一直传props
 const Context = React.createContext(null)
 
 function Navi(prop) {
+  const router = useRouterContext()
   return (
     <div className="dd-common-alignside" style={{ height: 45, padding: '0 15px', backgroundColor: 'white', borderBottom: '1px solid #eee' }}>
-      <div style={{ flex: 1, display: 'flex', flexDirection: 'row', }} onClick={() => { prop.router.back() }}>
+      <div style={{ flex: 1, display: 'flex', flexDirection: 'row', }} onClick={() => { router.back() }}>
         <div style={{ display: prop.showBack === false ? 'none' : 'flex' }}>
           <MIconView type="FaChevronLeft" />
         </div>