+++ src/loader/UserLoader.js
@@ -3,5 +3,5 @@ import UserModel from '../models/UserModel'
 import { createItemLoader } from './BaseLoader'
 
 export default createItemLoader(UserModel, async (params) => {
-    return services.getUserInfo(params)
+  return services.getUserInfo(params)
 })