+++ src/pages/UserShelfPage/index.js
@@ -3,13 +3,14 @@ import { useEffectOnce } from 'react-use'
 import { Button } from 'antd-mobile'
 import { Observer, useLocalStore } from 'mobx-react-lite'
 
-import { useRouterContext, useStoreContext } from 'contexts'
+import { useRouterContext, useStoreContext, useNaviContext } from 'contexts'
 import { LoaderListView, AutoCenterView } from 'components'
 import BookItem from 'business/ResourceItem/BookItem'
 
 
 export default function () {
   const globalStore = useStoreContext()
+  const Navi = useNaviContext()
   const router = useRouterContext()
   const loader = globalStore.bookShelfLoader
   const localStore = useLocalStore(() => ({
@@ -24,6 +25,7 @@ export default function () {
   return <Observer>{
     () => {
       return <Fragment>
+        <Navi title="书架"/>
         <LoaderListView
           loader={loader}
           renderEmpty={(